% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_stock.R
\name{adjusted_stock}
\alias{adjusted_stock}
\title{Adjusted stock}
\usage{
adjusted_stock(block_height, total_amount, total_amount_height)
}
\arguments{
\item{block_height}{The (integer) block height. May be a vector.}

\item{total_amount}{The \code{total_amount} of spendable bitcoins, in satoshis, at a particular block height.}

\item{total_amount_height}{The block height at which the \code{total_amount} parameter applies.}
}
\value{
The adjusted Bitcoin stock as an integer number of satoshis.
}
\description{
Returns the adjusted Bitcoin stock, in satoshis, for a given block height.
}
\examples{
adjusted_stock(1, 1896164634950948, 723896)
adjusted_stock(209999, 1896164634950948, 723896)
adjusted_stock(723896, 1896164634950948, 723896)

adjusted_stock(0:10, 1896164634950948, 723896)
}
