% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gross_stock.R
\name{gross_stock}
\alias{gross_stock}
\title{Gross stock}
\usage{
gross_stock(block_height)
}
\arguments{
\item{block_height}{The block height.}
}
\value{
The gross stock of bitcoins as an integer number of satoshis.
}
\description{
Returns the gross stock of bitcoins for a given block height.
}
\examples{
gross_stock(1)
gross_stock(100000)
gross_stock(723896)
}
\seealso{
\code{\link{execute_synthetic_experiment}}
}
