% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gross_stock.R
\name{gross_stock}
\alias{gross_stock}
\title{Gross stock}
\usage{
gross_stock(block_height, by_halving = FALSE)
}
\arguments{
\item{block_height}{The (integer) block height.}

\item{by_halving}{A logical value. If TRUE, the gross stock is returned as a vector
with one element per halving period.}
}
\value{
The gross Bitcoin stock as an integer number of satoshis.
}
\description{
Returns the gross Bitcoin stock, in satoshis, for a given block height.
}
\examples{
gross_stock(1)
gross_stock(209999)
gross_stock(210000)
gross_stock(210000, by_halving = TRUE)
gross_stock(420000, by_halving = TRUE)
gross_stock(723896)

gross_stock(0:10)
}
